Redid Searchbar Page
