+++ frontend/src/App.js
@@ -4,8 +4,6 @@ import DashLayout from './components/dash/dashlayout'
 import Banlists from './features/pages/banlist';
 import ContactInfos from './features/pages/contactinfo';
 import HomePage from './features/pages/homepage/homepage.js';
-import SearchResult from './features/pages/searchresult';
-import SearchResultLoggedIn from './features/pages/searchresultpage/SearchPageResultLoggedin.jsx';
 import AccountCreationForm from './features/pages/sign-up-page/AccountCreationForm.js'
 import LoginPage from './features/pages/loginpage/login.js';
 import Prefetch from './features/auth/Prefetch.js'
@@ -15,7 +13,7 @@ import MyDeck from './features/pages/mydeckpage/my-deck.js';
 import CreateNewDeckForm from './features/pages/mydeckpage/new-deck-creation-form.jsx';
 import Deckview from './features/pages/mydeckpage/modify-deck.jsx';
 import UserOwnedCardTable from './features/pages/my-cards/ownedCardTable.jsx';
-import SearchBar from './features/searchbar/searchbar.jsx';
+import SearchBar from './features/pages/searchpage/searchbar.jsx';
 
 
 function App() {
@@ -28,15 +26,11 @@ function App() {
           <Route path="banlist" element={<Banlists />} />
           <Route path="privacy-policy" element={<HomePage/>} />
           <Route path="contact-info" element={<ContactInfos />} />
-          
-          <Route path="search" element={<SearchBar/>}/>
-            <Route path=":cardname" element={<SearchResult />} />
-          </Route>
+          <Route path="search" element={<SearchBar/>} />
+        </Route>
 
           <Route element={<StayLoggedIn/>}>
             <Route element={<Prefetch/>}>
-              <Route path="searchloggedin" element={<SearchBar/>}/>
-              <Route path=":cardname" element={<SearchResultLoggedIn />}/>
               <Route path="card" element={<DashLayout />}>
                 <Route path="getcards" element={<UserOwnedCardTable/>}/>
               </Route>
