+++ frontend/src/components/header/header.js
@@ -4,7 +4,6 @@ import { Link, useNavigate, useLocation } from 'react-router-dom'
 import { faRightFromBracket } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import Cardsearch from '../buttons/searchbar'
-import CardsearchLoggedIn from '../buttons/searchbarloggedin'
 import Banlist from '../buttons/banlist'
 import Mycards from '../buttons/my-cards'
 import Login from '../buttons/login'
@@ -47,10 +46,7 @@ const Header = () => {
     }
 
     const renderCardSearchButton = () => {
-        if (!isAuthenticated) {
-            return <Cardsearch />;
-        } 
-        return <CardsearchLoggedIn />;
+        return <Cardsearch />;
     }
 
     const renderAuthButtons = () => {
