+++ frontend/src/features/pages/searchpage/styling/searchbar.css
@@ -97,25 +97,32 @@ input:focus {
   flex-direction: column;
   align-items: center;
   width: 70%;
-  max-height: 64vh;
+  justify-content: center;
+  min-height: 30vh;
+  max-height: 60vh;
   margin-top: 5%;
   margin-bottom: 3%;
+  overflow: hidden;
 }
 
 .Searchbar-suggestion-grid {
-  display: Grid;
-  grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
   gap: 25px;
   position: relative;
-  overflow-y: auto;
   width: 100%;
-  position: relative;
+  max-width: 100%;
+  max-height: 100%;
+  justify-items: stretch; 
   overflow-y: auto;
+  justify-items: center;
 }
+
 
 .gallery-item-image {
   width: 100%;
-  position: relative;
+  max-width: 550px;
+  height: auto; 
 }
 
 /* Pagination */
