--- frontend/src/features/pages/searchresultpage/styling/SearchResult.css
@@ -1,325 +1,0 @@
-.logged-in-search-result-page-body {
-    background-color: transparent;
-    position: relative;
-    display: grid;
-    place-items: center;
-    min-height: 82vh;
-    padding: 20px;
-    background-color: rgb(32, 32, 32);
-    background-image: linear-gradient(45deg, black 25%, transparent 25%, transparent 75%, black 75%, black), linear-gradient(45deg, black 25%, transparent 25%, transparent 75%, black 75%, black), linear-gradient(to bottom, rgb(8, 8, 8), rgb(32, 32, 32));
-    background-size: 10px 10px, 10px 10px, 10px 5px;
-    background-position: 0px 0px, 5px 5px, 0px 0px;
-}
-
-.card-background {
-    background-color: rgb(32, 32, 32);
-    background-image: linear-gradient(45deg, black 25%, transparent 25%, transparent 75%, black 75%, black), linear-gradient(45deg, black 25%, transparent 25%, transparent 75%, black 75%, black), linear-gradient(to bottom, rgb(8, 8, 8), rgb(32, 32, 32));
-    background-size: 10px 10px, 10px 10px, 10px 5px;
-    background-position: 0px 0px, 5px 5px, 0px 0px;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    min-height: 75vh;
-    height: auto;
-    padding: 6%;
-}
-
-.Card-information-container {
-    flex-grow: 1;
-    height: fit-content;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: -5%;
-}
-
-.Search-Result-back-icon-container {
-    position: relative;
-    z-index: 10;
-    margin-top: -5%;
-    right: -4%
-}
-
-.Search-Result-back-icon {
-    background-color: transparent;
-    border: none;
-    color: gold;
-    cursor: pointer;
-}
-
-.search-card-name {
-    color: goldenrod;
-    text-align: center;
-    height: fit-content;
-    background: rgba(95, 94, 94, 0.363);    
-    border-radius: 15px;
-    padding: 10px;
-    width: 40%;
-    margin-bottom: 2%;
-    font-size: calc(15px + 0.390625vw);
-    font-weight: bold;
-}
-
-
-.cardbody {
-    display: flex;
-    flex-direction: row;
-    width: 80%;
-}
-
-.normal-desc-container,
-.desc-container,
-.monster-desc-container,
-.pend-desc-container {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-}
-
-.normal-desc-header,
-.monster-desc-header,
-.pend-desc-header  {
-    display: flex;
-    color: darkorange;
-    align-items: center;
-    padding-left: 4%;
-    padding-right: 4%;
-    width: 20%;
-    font-size: calc(8px + 0.390625vw);
-}
-
-.normal-desc-body,
-.monster-desc-body,
-.pend-desc-body  {
-    background: rgba(95, 94, 94, 0.185);
-    padding-left: 6%;
-    padding-right: 6%;
-    padding-top: 2%;
-    padding-bottom: 2%;
-    color: goldenrod;
-    width: 80%;
-    border-radius: 25px;
-    font-size: calc(8px + 0.390625vw);
-}
-
-.card-type-container,
-.card-subtype-container,
-.card-attribute-container,
-.card-archetype-container,
-.card-level-container,
-.card-pendulum-scale-container,
-.card-link-container,
-.card-atk-container,
-.card-def-container {
-    display: flex;
-    flex-direction: row;
-    background-color: transparent;
-    width: 100%;
-    margin-bottom: 5%;
-
-}
-
-.card-type-header,
-.card-subtype-header,
-.card-attribute-header,
-.card-archetype-header,
-.card-level-header,
-.card-pendulum-scale-header,
-.card-link-header,
-.card-atk-header,
-.card-def-header {
-    background: linear-gradient(to right, rgb(31, 27, 27), rgb(95, 94, 94));
-    display: flex;
-    color: darkorange;
-    padding-left: 4%;
-    padding-right: 4%;
-    padding-top: 2%;
-    padding-bottom: 2%;
-    width: 33%;
-    font-size: calc(8px + 0.390625vw);
-}
-
-.card-subtype-body,
-.card-type-body,
-.card-attribute-body,
-.card-archetype-body,
-.card-level-body,
-.card-pendulum-scale-body,
-.card-link-body,
-.card-atk-body,
-.card-def-body {
-    padding-left: 6%;
-    padding-right: 6%;
-    padding-top: 2%;
-    padding-bottom: 2%;
-    color: goldenrod;
-    width: 50%;
-    font-size: calc(8px + 0.390625vw);
-}
-
-.card-level,
-.card-link,
-.card-attribute,
-.card-archetype{
-    background: linear-gradient(to left, rgb(31, 27, 27), rgb(95, 94, 94));;
-    color: goldenrod;
-    text-align: center;
-    font-size: calc(11px + 0.390625vw);
-    font-weight: bold;
-    position: relative;
-    width: 100%;
-    height: fit-content;
-    padding-left: 6%;
-    padding-right: 6%;
-    padding-top: 2%;
-    padding-bottom: 2%;
-    margin-bottom: 4%;
-    margin-left: 0.5%;
-    border-radius: 10px;    
-}
-
-
-
-.card-descriptions-container {
-    display:flex;
-    position: relative;
-    flex-direction: column;
-    color: goldenrod;
-    padding: 10px;
-    height: fit-content;
-    width: 80%
-}
-
-.cardstats {
-    display: flex;
-    flex-direction: column;
-}
-
-.button-container {
-    display: flex;
-    justify-content: center;
-    width: fit-content;
-    margin-top: 20px;
-    width: 100%;
-}
-
-.my-button {
-    padding: 10px 20px;
-    border-radius: 15px;
-    font-size: 16px;
-    cursor: pointer;
-    background: linear-gradient(to left, rgb(31, 27, 27), rgb(95, 94, 94));;
-    color: goldenrod;
-    text-align: center;
-}
-
-.success-message {
-    color: green;
-    margin-left: 10px;
-    font-weight: bold;
-}
-
-.columnone {
-    
-    display: flex;
-    flex-direction: column;
-    padding: 1%;
-    margin-bottom: 3%;
-    width: 50%;
-    height: fit-content;
-}
-
-
-
-.columntwo {
-    display: flex;
-    flex-direction: column;
-    padding: 1%;
-    margin-bottom: 3%;
-    margin-left: 1%;
-    width: 50%;
-    height: fit-content;
-}
-
-.Card-container {
-    display: flex;
-    background: linear-gradient(to bottom, rgb(45, 43, 43), black);
-    width: 100%;
-    padding: 6%;
-    border-radius: 25px;
-}
-
-.logged-in-card-container {
-    display: flex;
-    background: linear-gradient(to bottom, rgb(45, 43, 43), black);
-    width: 90%;
-    padding: 6%;
-    border-radius: 25px;
-}
-
-
-.card-image {
-    width: 300px;
-    min-height: 450px;
-    object-fit: contain;
-    margin-right: 20px;
-}
-
-.card-details {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    margin-left: 13%;
-    padding: 2%;
-    width: 100%
-}
-
-/*loading styling*/
-.loading-spinner-icon-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 7%;
-}
-
-.logged-in-loading-spinner-icon-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: -2%
-}
-
-.loading-spinner-icon {
-    color: goldenrod;
-    margin-right: 70px;
-}
-
-.loading-spinner-text {
-    font-size: 50px;
-    color: white
-}
-
-/* mobile devices */
-@media screen and (width <= 480px) {
-}
-  
-/* medium Devices (iPads, Landscape tablets) */
-@media screen and (481px <= width <= 768px) {
-}
-
-/* medium-large devices (laptops and small screen) */
-@media screen and (769px <= width <= 1024px) {
-}
-  
-/* Large devices (laptops/desktops) */
-@media screen and (1025px <= width <= 1200px) {
-}
-  
-/* Extra large laptops */
-@media screen and (1201px <= width <= 1536px) {
-}
-  
-/* TV and large desktops */
-@media screen and (1537px <= width) { 
-}
